+++ main.tf
@@ -49,6 +49,10 @@ data "template_file" "nomad_job_minio" {
     envs                      = local.minio_env_vars
     use_host_volume           = var.use_host_volume
     use_canary                = var.use_canary
+    use_vault_kms             = var.use_vault_kms
+    vault_address             = var.vault_address
+    vault_kms_approle_kv      = var.vault_kms_approle_kv
+    vault_kms_key_name        = var.vault_kms_key_name
     upstreams                 = jsonencode(var.minio_upstreams)
     cpu_proxy                 = var.resource_proxy.cpu
     memory_proxy              = var.resource_proxy.memory